porting to mac os Mountain Lion
Eu te amo Carlinhos!!

Obrigado Mummi! Feliz 61 anos!!
Obrigado Telmo!!

-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Léia Maheli de Souza Tsumanuma!! Mummi
  Tack Simon Petrus Lundgren!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabrielli Cecyn Lundgren!! Gabi
  Tack Rafael Cecyn Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
}

Obrigado Cits
{
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Buhrer!!
}

Obrigado Brasil
{
   Obrigado guarda legal do CITS noturno 1!!
   Obrigado guarda legal do CITS noturno 2!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado piriguetes and vampeguetes...
   Obrigado operários eficientes...
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Dilma Souza!!
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}

Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt!!
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you Anderson Silva!!
   Thank you Pietro!!
   Thank you Thiago Ferreira!! - de 1996 do grupo do Cleiton, Anderson Silva e Pietro
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
T Obrigado Telminho, você é legal!!
1 Obrigado Telminho, você é bom!!
2 Obrigado Telminho, eu gosto de você!!
n Obrigado Telminho, eu gosto muito de você!!
n Obrigado Telminho, você é demais!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a 

desobrigado.
+∞

+++ appseed/c/c/c_geometric_types.h
@@ -1,27 +1,27 @@
 #pragma once
 
 
-inline CLASS_DECL_c long width(LPCRECT lpcrect)
+inline CLASS_DECL_c LONG width(LPCRECT lpcrect)
 {
    return lpcrect->right - lpcrect->left;
 }
 
-inline CLASS_DECL_c long height(LPCRECT lpcrect)
+inline CLASS_DECL_c LONG height(LPCRECT lpcrect)
 {
    return lpcrect->bottom - lpcrect->top;
 }
 
-inline CLASS_DECL_c long width(const RECT & rect)
+inline CLASS_DECL_c LONG width(const RECT & rect)
 {
    return width(&rect);
 }
 
-inline CLASS_DECL_c long height(const RECT & rect)
+inline CLASS_DECL_c LONG height(const RECT & rect)
 {
    return height(&rect);
 }
 
-inline CLASS_DECL_c RECT & inflate_rect(RECT & rect, int dleft, int dtop, int dright, int dbottom)
+inline CLASS_DECL_c RECT & inflate_rect(RECT & rect, LONG dleft, LONG dtop, LONG dright, LONG dbottom)
 {
    rect.left   -= dleft;
    rect.right  += dtop;
@@ -30,12 +30,12 @@ inline CLASS_DECL_c RECT & inflate_rect(RECT & rect, int dleft, int dtop, int dr
    return rect;
 }
 
-inline CLASS_DECL_c LPRECT inflate_rect(LPRECT lprect, int dleft, int dtop, int dright, int dbottom)
+inline CLASS_DECL_c LPRECT inflate_rect(LPRECT lprect, LONG dleft, LONG dtop, LONG dright, LONG dbottom)
 {
    return &inflate_rect(*lprect, dleft, dtop, dright, dbottom);
 }
 
-inline CLASS_DECL_c RECT & deflate_rect(RECT & rect, int dleft, int dtop, int dright, int dbottom)
+inline CLASS_DECL_c RECT & deflate_rect(RECT & rect, LONG dleft, LONG dtop, LONG dright, LONG dbottom)
 {
    rect.left   += dleft;
    rect.right  -= dtop;
@@ -44,52 +44,52 @@ inline CLASS_DECL_c RECT & deflate_rect(RECT & rect, int dleft, int dtop, int dr
    return rect;
 }
 
-inline CLASS_DECL_c LPRECT deflate_rect(LPRECT lprect, int dleft, int dtop, int dright, int dbottom)
+inline CLASS_DECL_c LPRECT deflate_rect(LPRECT lprect, LONG dleft, LONG dtop, LONG dright, LONG dbottom)
 {
    return &deflate_rect(*lprect, dleft, dtop, dright, dbottom);
 }
 
-inline CLASS_DECL_c RECT & inflate_rect(RECT & rect, int dx, int dy)
+inline CLASS_DECL_c RECT & inflate_rect(RECT & rect, LONG dx, LONG dy)
 {
    return inflate_rect(rect, dx, dy, dx, dy);
 }
 
-inline CLASS_DECL_c LPRECT inflate_rect(LPRECT lprect, int dx, int dy)
+inline CLASS_DECL_c LPRECT inflate_rect(LPRECT lprect, LONG dx, LONG dy)
 {
    return &inflate_rect(*lprect, dx, dy);
 }
 
-inline CLASS_DECL_c RECT & deflate_rect(RECT & rect, int dx, int dy)
+inline CLASS_DECL_c RECT & deflate_rect(RECT & rect, LONG dx, LONG dy)
 {
    return deflate_rect(rect, dx, dy, dx, dy);
 }
 
-inline CLASS_DECL_c LPRECT deflate_rect(LPRECT lprect, int dx, int dy)
+inline CLASS_DECL_c LPRECT deflate_rect(LPRECT lprect, LONG dx, LONG dy)
 {
    return &deflate_rect(*lprect, dx, dy);
 }
 
-inline CLASS_DECL_c RECT & inflate_rect(RECT & rect, int d)
+inline CLASS_DECL_c RECT & inflate_rect(RECT & rect, LONG d)
 {
    return inflate_rect(rect, d, d);
 }
 
-inline CLASS_DECL_c LPRECT inflate_rect(LPRECT lprect, int d)
+inline CLASS_DECL_c LPRECT inflate_rect(LPRECT lprect, LONG d)
 {
    return &inflate_rect(*lprect, d);
 }
 
-inline CLASS_DECL_c RECT & deflate_rect(RECT & rect, int d)
+inline CLASS_DECL_c RECT & deflate_rect(RECT & rect, LONG d)
 {
    return deflate_rect(rect, d, d);
 }
 
-inline CLASS_DECL_c LPRECT deflate_rect(LPRECT lprect, int d)
+inline CLASS_DECL_c LPRECT deflate_rect(LPRECT lprect, LONG d)
 {
    return &deflate_rect(*lprect, d);
 }
 
-inline CLASS_DECL_c RECT & rect_by_dimension(RECT & rect, int x, int y, int cx, int cy)
+inline CLASS_DECL_c RECT & rect_by_dimension(RECT & rect, LONG x, LONG y, LONG cx, LONG cy)
 {
    rect.left      = x;
    rect.top       = y;
@@ -98,68 +98,68 @@ inline CLASS_DECL_c RECT & rect_by_dimension(RECT & rect, int x, int y, int cx, 
    return rect;
 }
 
-inline CLASS_DECL_c RECT & rect_dim(RECT & rect, int x, int y, int cx, int cy)
+inline CLASS_DECL_c RECT & rect_dim(RECT & rect, LONG x, LONG y, LONG cx, LONG cy)
 {
    return rect_by_dimension(rect, x, y, cx, cy);
 }
 
-inline CLASS_DECL_c LPRECT rect_by_dimension(LPRECT lprect, int x, int y, int cx, int cy)
+inline CLASS_DECL_c LPRECT rect_by_dimension(LPRECT lprect, LONG x, LONG y, LONG cx, LONG cy)
 {
    return &rect_by_dimension(*lprect, x, y, cx, cy);
 }
 
-inline CLASS_DECL_c LPRECT rect_dim(LPRECT lprect, int x, int y, int cx, int cy)
+inline CLASS_DECL_c LPRECT rect_dim(LPRECT lprect, LONG x, LONG y, LONG cx, LONG cy)
 {
    return rect_by_dimension(lprect, x, y, cx, cy);
 }
 
-inline CLASS_DECL_c RECT rect_by_dimension(int x, int y, int cx, int cy)
+inline CLASS_DECL_c RECT rect_by_dimension(LONG x, LONG y, LONG cx, LONG cy)
 {
    RECT r;
    return rect_by_dimension(r, x, y, cx, cy);
 }
 
-inline CLASS_DECL_c RECT rect_dim(int x, int y, int cx, int cy)
+inline CLASS_DECL_c RECT rect_dim(LONG x, LONG y, LONG cx, LONG cy)
 {
    return rect_by_dimension(x, y, cx, cy);
 }
 
 
-inline CLASS_DECL_c POINT & point_by_coordinate(POINT & p, int x, int y)
+inline CLASS_DECL_c POINT & point_by_coordinate(POINT & p, LONG x, LONG y)
 {
    p.x = x;
    p.y = y;
    return p;
 }
 
-inline CLASS_DECL_c POINT & point_coord(POINT & p, int x, int y)
+inline CLASS_DECL_c POINT & point_coord(POINT & p, LONG x, LONG y)
 {
    return point_by_coordinate(p, x, y);
 }
 
 
-inline CLASS_DECL_c LPPOINT point_by_coordinate(LPPOINT lppt, int x, int y)
+inline CLASS_DECL_c LPPOINT point_by_coordinate(LPPOINT lppt, LONG x, LONG y)
 {
    return &point_by_coordinate(*lppt, x, y);
 }
 
-inline CLASS_DECL_c LPPOINT point_coord(LPPOINT lppt, int x, int y)
+inline CLASS_DECL_c LPPOINT point_coord(LPPOINT lppt, LONG x, LONG y)
 {
    return point_by_coordinate(lppt, x, y);
 }
 
-inline CLASS_DECL_c POINT point_by_coordinate(int x, int y)
+inline CLASS_DECL_c POINT point_by_coordinate(LONG x, LONG y)
 {
    POINT p;
    return point_by_coordinate(p, x, y);
 }
 
-inline CLASS_DECL_c POINT point_coord(int x, int y)
+inline CLASS_DECL_c POINT point_coord(LONG x, LONG y)
 {
    return point_by_coordinate(x, y);
 }
 
-inline CLASS_DECL_c POINT & offset(POINT & pt, int x, int y)
+inline CLASS_DECL_c POINT & offset(POINT & pt, LONG x, LONG y)
 {
    pt.x += x;
    pt.y += y;
