+++ appseed/c/simple_ui/simple_label.cpp
@@ -46,7 +46,7 @@ void simple_label::draw_this(simple_graphics & g)
 
    simple_font f;
 
-   f.create_pixel(g, height(m_rect) * 10, "Geneva");
+   f.create_pixel(g, (int)height(m_rect) * 10, "Geneva");
 
    float fMargin = (height(m_rect) * ((1.0f - 0.7f) / 2.0f));
