+++ appseed/c/simple_ui/simple_tap.cpp
@@ -208,11 +208,11 @@ void simple_tap::draw_volume(simple_graphics & g)
 
       simple_linear_gradient_brush br1(g, m_rect.left, m_rect.top - 1, m_rect.left, m_rect.top + iBorderH + 2, crOut, crIn);
 
-      g.fill_rect(rect_dim(m_rect.left, m_rect.top, width(&m_rect), iBorderH), br1);
+      g.fill_rect(rect_dim(m_rect.left, m_rect.top, (int) width(&m_rect), iBorderH), br1);
 
       simple_linear_gradient_brush br2(g, m_rect.left, m_rect.top + iBorderH - 1, m_rect.left, m_rect.top + iBorderH * 2 + 2, crIn, crOut);
 
-      g.fill_rect(rect_dim( m_rect.left, m_rect.top + iBorderH, width(&m_rect), iBorderH), br2);
+      g.fill_rect(rect_dim( m_rect.left, m_rect.top + iBorderH, (int) width(&m_rect), iBorderH), br2);
 
       /*Gdiplus::Pen pen1(crBorderOut);
 
@@ -222,7 +222,7 @@ void simple_tap::draw_volume(simple_graphics & g)
 
       simple_solid_pen pen(g, crBorderIn);
 
-      g.draw_rect(rect_dim(m_rect.left + 1, m_rect.top + 1, width(&m_rect) - 2, iBorderH * 2 - 2), pen);
+      g.draw_rect(rect_dim(m_rect.left + 1, m_rect.top + 1, (int) width(&m_rect) - 2, iBorderH * 2 - 2), pen);
 
    }
 
@@ -248,7 +248,7 @@ void simple_tap::draw_text(simple_graphics & g)
 
    g.select(b);
 
-   simple_pixel_font f(g, height(m_rect) * 10, "Geneva");
+   simple_pixel_font f(g, (int) height(m_rect) * 10, "Geneva");
 
    g.select(f);
